NEON-1933 Fix count by column width
If the aggregation table (aka "count by" widget) initializes before it has the
proper size, then the columns will be squished to minimum size. The columns need
to be resized when the visualization resizes. The ag-Grid resizes the grid every
500ms, and doesn't provide notification of when it does so (the gridSizeChanged
event never ever fires). So instead, manually tell ag-Grid to resize when the
visualization resizes.