Add applyTransientDateFilter option to the map layer configs
If this settings is false (the default), then the map will not respond to the
date-selected events that the timeline and line chart send out when the user
mouses over dates. When it is true, the corresponding layer will only show data
from the moused-over date. This can be useful, but it can also have a large
performance hit.