Fix error handling of translation in news feed
Google Translate has a maximum number of characters it will accept in requests, so make sure we
don't send more than that. If an error does occur in translation, indicate this to the user but
allow translations to continue. Also, fix the translation tooltips so that they match the item that
the user hovered over.